Move site-configs.ts to site-configs directory
