+++ site-configs/main.ts
@@ -1,4 +1,4 @@
-import { Environment, GetSiteConfig } from "../site-configs";
+import { Environment, GetSiteConfig } from "./site-configs";
 
 const envToDomainMap: Record<Environment, string> = {
     local: "localhost:3000",
